requirements.txt
# Core ML and Data Science
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.0
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.17.0

# Network and Security Analysis
scapy==2.5.0
netaddr==0.9.0
pyshark==0.6

# Data Processing
joblib==1.3.2
imbalanced-learn==0.11.0

# SIEM Integration
elasticsearch==8.11.0
kafka-python==2.0.2
redis==5.0.1

# Web Framework (for dashboard)
flask==3.0.0
flask-socketio==5.3.6
dash==2.14.2
dash-bootstrap-components==1.5.0

# Monitoring and Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Configuration and Utils
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0

# Development
jupyter==1.0.0
pytest==7.4.3
black==23.11.0
flake8==6.1.0
